+++ src/SignalDecomposition.jl
@@ -27,6 +27,9 @@ include("product/matrixinversion.jl")
 include("nonlinear/extremelysimple.jl")
 include("nonlinear/projection.jl")
 include("misc/anomaly.jl")
+include("detrending/simple.jl")
+include("detrending/smoothing.jl")
+
 
 
 end # module